Refactored and moved to new file
